refactor(worker): move internal/worker/{pool,executor,status} → worker/pool
Pool, executor, and the internal statusAccumulator relocate with package
rename (worker → pool). The prompt builder stays behind temporarily
(moved in Task 15). History preserved via git mv.

status.go accompanied pool.go and executor.go because it's the same
package — the plan spec listed only pool/executor but status.go is
referenced by pool.go's reportStatus helpers.

Cross-module wiring: worker/go.mod now requires+replaces the root module
so worker/pool can import internal/bot (for bot.RunOptions) and
internal/worker (for AttachmentInfo/BuildPrompt, until Task 15 extracts
the prompt builder into worker/prompt).

Callers in cmd/agentdock/{local_adapter,worker}.go used the `pool`
identifier for local variables/fields, so the import is aliased as
`workerpool` there to keep field/var names intact (Task 13 moves
local_adapter.go into worker/pool anyway). Integration tests renamed
their local `pool` variables to `p` instead.
