+++ worker/pool/status.go
@@ -1,4 +1,4 @@
-package worker
+package pool
 
 import (
 	"sync"
